vo_conesearch: Renamed async to vo_async as async is reserved keyword stating with Python 3.7
